requirements.txt
beautifulsoup4==4.11.1
bottle==0.12.21
bottle-websocket==0.2.9
branca==0.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Eel==0.14.0
flake8==4.0.1
Flask==2.1.2
folium==0.12.1.post1
future==0.18.2
geojson==2.5.0
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.23.0
Pillow==9.2.0
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pyfiglet==0.8.post1
pyflakes==2.4.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyowm==3.3.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==0.20.0
pytweening==1.0.4
pywhatkit==5.4
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.9
Werkzeug==2.1.2
whichcraft==0.6.1
wikipedia==1.4.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
